color: move color_dump()
Move color_dump() to color/debug.c where it belongs.
