+++ color/curses2.h
@@ -46,11 +46,6 @@ struct CursesColor
 };
 TAILQ_HEAD(CursesColorList, CursesColor);
 
-#ifdef USE_DEBUG_COLOR
-extern struct CursesColorList CursesColors;
-extern int NumCursesColors;
-#endif
-
 void                curses_color_free(struct CursesColor **ptr);
 struct CursesColor *curses_color_new (int fg, int bg);
